requirements.txt
feedparser==6.0.11
html2text==2020.1.16
python-dotenv==1.0.0
Requests==2.31.0
